requirements.txt
flask==0.12
gensim==1.0.1
kafka-python==1.3.3
nltk==3.2.2
numpy==1.12.1
redis==3.2.0
tweepy==3.5.0
Werkzeug==0.12
psutil==5.2.0
